+++ CHANGELOG.rst
@@ -7,6 +7,12 @@ New features
 ~~~~~~~~~~~~
 - MAS output files with timestamps > 3 digits long can now be read in.
 
+New behaviour
+~~~~~~~~~~~~~
+- `psipy.model.variable.Variable.sample_at_coords` will now return ``NaN`` for
+  any sample points that are out of bounds, and raise a warning letting
+  the user know some points were out of bounds.
+
 Breaking changes
 ~~~~~~~~~~~~~~~~
 - The radial coordinate input to `psipy.tracing.FortranTracer.trace` is now required to be